+++ src/sas/sascalc/shape2sas/Typing.py
@@ -1,0 +1,6 @@
+import numpy as np
+from typing import Optional, Tuple, List, Any
+
+Vector2D = Tuple[np.ndarray, np.ndarray]
+Vector3D = Tuple[np.ndarray, np.ndarray, np.ndarray]
+Vector4D = Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]
