+++ src/sas/sascalc/shape2sas/helpfunctions.py
@@ -1,15 +1,7 @@
 import numpy as np
 import matplotlib.pyplot as plt
-from scipy.special import gamma
-from typing import Optional, Tuple, List, Any
-#from dataclasses import dataclass
-
-
-################################ Type Hints ################################
-Vector2D = Tuple[np.ndarray, np.ndarray]
-Vector3D = Tuple[np.ndarray, np.ndarray, np.ndarray]
-Vector4D = Tuple[np.ndarray, np.ndarray, np.ndarray, np.ndarray]
-
+from sas.sascalc.shape2sas.Typing import *
+from sas.sascalc.shape2sas.models import *
 
 ################################ Shape2SAS helper functions ###################################
 def sinc(x) -> np.ndarray:
@@ -18,493 +10,6 @@ def sinc(x) -> np.ndarray:
     numpy.sinc is defined as sinc(x) = sin(pi*x)/(pi*x)
     """
     return np.sinc(x / np.pi)   
-
-'''#template to write a subunit class
-class <NAME OF SUBUNIT HERE>:
-    def __init__(self, dimensions: List[float]):
-        #PARAMERERS HERE
-        self.<PARAMETER NAME> = dimensions[0]
-        self.<PARAMETER NAME> = dimensions[1]
-        self.<PARAMETER NAME> = dimensions[2]
-    
-    def getVolume(self) -> float:
-        """Returns the volume of the subunit"""
-
-        <WRITE VOLUME CALCULATION OF THE SUBUNIT HERE>
-
-        return <VOLUME>
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of the subunit"""
-
-        Volume = self.getVolume()
-        Volume_max = <MAXIMUM VOLUME> ###Box around the subunit
-        Vratio = Volume_max/Volume
-
-        N = int(Vratio * Npoints)
-
-        <WRITE POINT DISTRIBUTION HERE>
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray,
-                            y_eff: np.ndarray,
-                            z_eff: np.ndarray) -> np.ndarray:
-          """Check for points within the subunit"""
-    
-          <WRITE OVERLAP CHECK HERE>
-    
-          return idx
-'''
-
-
-class Sphere:
-    def __init__(self, dimensions: List[float]):
-        self.R = dimensions[0]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a sphere"""
-        return (4 / 3) * np.pi * self.R**3
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a sphere"""
-        Volume = self.getVolume()
-        Volume_max = (2*self.R)**3 ###Box around sphere.
-        Vratio = Volume_max/Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.R, self.R, N)
-        y = np.random.uniform(-self.R, self.R, N)
-        z = np.random.uniform(-self.R, self.R, N)
-        d = np.sqrt(x**2 + y**2 + z**2)
-
-        idx = np.where(d < self.R) #save points inside sphere
-        x_add,y_add,z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, 
-                     x_eff: np.ndarray, 
-                     y_eff: np.ndarray, 
-                     z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a sphere"""
-
-        d = np.sqrt(x_eff**2+y_eff**2+z_eff**2)
-        idx = np.where(d > self.R)
-        return idx
-
-    
-class HollowSphere:
-    def __init__(self, dimensions: List[float]):
-        self.R = dimensions[0]
-        self.r = dimensions[1]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a hollow sphere"""
-        if self.r > self.R:
-            self.R, self.r = self.r, self.R
-
-        if self.r == self.R:
-            return 4 * np.pi * self.R**2 #surface area of a sphere
-        else: 
-            return (4 / 3) * np.pi * (self.R**3 - self.r**3)
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a hollow sphere"""
-        Volume = self.getVolume()
-
-        if self.r == self.R:
-            #The hollow sphere is a shell
-            phi = np.random.uniform(0,2 * np.pi, Npoints)
-            costheta = np.random.uniform(-1, 1, Npoints)
-            theta = np.arccos(costheta)
-
-            x_add = self.R * np.sin(theta) * np.cos(phi)
-            y_add = self.R * np.sin(theta) * np.sin(phi)
-            z_add = self.R * np.cos(theta)
-            return x_add, y_add, z_add
-
-        Volume_max = (2*self.R)**3 ###Box around the sphere
-        Vratio = Volume_max/Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.R, self.R, N)
-        y = np.random.uniform(-self.R, self.R, N)
-        z = np.random.uniform(-self.R, self.R, N)
-        d = np.sqrt(x**2 + y**2 + z**2)
-
-        idx = np.where((d < self.R) & (d > self.r))
-        x_add, y_add, z_add = x[idx], y[idx], z[idx]
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a hollow sphere"""
-
-        d = np.sqrt(x_eff**2+y_eff**2+z_eff**2)
-        if self.r > self.R:
-             self.r, self.R = self.R, self.r
-
-        if self.r == self.R:
-            idx = np.where(d != self.R)
-            return idx
-        
-        else:
-            idx = np.where((d > self.R) | (d < self.r))
-            return idx
-
-
-class Cylinder:
-    def __init__(self, dimensions: List[float]):
-        self.R = dimensions[0]
-        self.l = dimensions[1]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a cylinder"""
-        return np.pi * self.R**2 * self.l
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a cylinder"""
-        Volume = self.getVolume()
-        Volume_max = 2 * self.R * 2 * self.R * self.l
-        Vratio = Volume_max / Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.R, self.R, N)
-        y = np.random.uniform(-self.R, self.R, N)
-        z = np.random.uniform(-self.l / 2, self.l / 2, N)
-        d = np.sqrt(x**2 + y**2)
-        idx = np.where(d < self.R)
-        x_add,y_add,z_add = x[idx],y[idx],z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a cylinder"""
-        d = np.sqrt(x_eff**2+y_eff**2)
-        idx = np.where((d > self.R) | (abs(z_eff) > self.l / 2))
-        return idx
-
-
-class Ellipsoid:
-    def __init__(self, dimensions: List[float]):
-        self.a = dimensions[0]
-        self.b = dimensions[1]
-        self.c = dimensions[2]
-
-    def getVolume(self) -> float:
-        """Returns the volume of an ellipsoid"""
-        return (4 / 3) * np.pi * self.a * self.b * self.c
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of an ellipsoid"""
-        Volume = self.getVolume()
-        Volume_max = 2 * self.a * 2 * self.b * 2 * self.c
-        Vratio = Volume_max / Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.a, self.a, N)
-        y = np.random.uniform(-self.b, self.b, N)
-        z = np.random.uniform(-self.c, self.c, N)
-
-        d2 = x**2 / self.a**2 + y**2 / self.b**2 + z**2 / self.c**2
-        idx = np.where(d2 < 1)
-        x_add, y_add, z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """check for points within a ellipsoid"""
-        d2 = x_eff**2 / self.a**2 + y_eff**2 / self.b**2 + z_eff**2 / self.c**2
-        idx = np.where(d2 > 1)
-        return idx
-
-
-class EllipticalCylinder:
-    def __init__(self, dimensions: List[float]):
-        self.a = dimensions[0]
-        self.b = dimensions[1]
-        self.l = dimensions[2]
-
-    def getVolume(self) -> float:
-        """Returns the volume of an elliptical cylinder"""
-        return np.pi * self.a * self.b * self.l
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of an elliptical cylinder"""
-        Volume = self.getVolume()
-        Volume_max = 2 * self.a * 2 * self.b * self.l
-        Vratio = Volume_max / Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.a, self.a, N)
-        y = np.random.uniform(-self.b, self.b, N)
-        z = np.random.uniform(-self.l / 2, self.l / 2, N)
-
-        d2 = x**2 / self.a**2 + y**2 / self.b**2
-        idx = np.where(d2 < 1)
-        x_add, y_add, z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a Elliptical cylinder"""
-        d2 = x_eff**2 / self.a**2 + y_eff**2 / self.b**2
-        idx = np.where((d2 > 1) | (abs(z_eff) > self.l / 2))
-        return idx
-
-
-class Disc(EllipticalCylinder):
-    pass
-
-
-class Cube:
-    def __init__(self, dimensions: List[float]):
-        self.a = dimensions[0]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a cube"""
-        return self.a**3
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a cube"""
-
-        #Volume = self.getVolume()
-        N = Npoints
-        x_add = np.random.uniform(-self.a / 2, self.a / 2, N)
-        y_add = np.random.uniform(-self.a / 2, self.a / 2, N)
-        z_add = np.random.uniform(-self.a / 2, self.a / 2, N)
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a cube"""
-        idx = np.where((abs(x_eff) >= self.a/2) | (abs(y_eff) >= self.a/2) | 
-            (abs(z_eff) >= self.a/2))
-        return idx
-
-
-class HollowCube:
-    def __init__(self, dimensions: List[float]):
-        self.a = dimensions[0]
-        self.b = dimensions[1]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a hollow cube"""
-
-        if self.a < self.b:
-            self.a, self.b = self.b, self.a
-
-        if self.a == self.b:
-            return 6 * self.a**2 #surface area of a cube
-
-        else: 
-            return (self.a - self.b)**3
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a hollow cube"""
-
-        Volume = self.getVolume()
-        
-        if self.a == self.b:
-            #The hollow cube is a shell
-            d = self.a / 2
-            N = int(Npoints / 6)
-            one = np.ones(N)
-            
-            #make each side of the cube at a time
-            x_add, y_add, z_add = [], [], []
-            for sign in [-1, 1]:
-                x_add = np.concatenate((x_add, sign * one * d))
-                y_add = np.concatenate((y_add, np.random.uniform(-d, d, N)))
-                z_add = np.concatenate((z_add, np.random.uniform(-d, d, N)))
-                
-                x_add = np.concatenate((x_add, np.random.uniform(-d, d, N)))
-                y_add = np.concatenate((y_add, sign * one * d))
-                z_add = np.concatenate((z_add, np.random.uniform(-d, d, N)))
-
-                x_add = np.concatenate((x_add, np.random.uniform(-d, d, N)))
-                y_add = np.concatenate((y_add, np.random.uniform(-d, d, N)))
-                z_add = np.concatenate((z_add, sign * one * d))
-            return x_add, y_add, z_add
-        
-        Volume_max = self.a**3
-        Vratio = Volume_max / Volume
-        N = int(Vratio * Npoints)
-
-        x = np.random.uniform(-self.a / 2,self.a / 2, N)
-        y = np.random.uniform(-self.a / 2,self.a / 2, N)
-        z = np.random.uniform(-self.a / 2,self.a / 2, N)
-
-        d = np.maximum.reduce([abs(x), abs(y), abs(z)])
-        idx = np.where(d >= self.b / 2)
-        x_add,y_add,z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a hollow cube"""
-
-        if self.a < self.b:
-            self.a, self.b = self.b, self.a
-        
-        if self.a == self.b:
-            idx = np.where((abs(x_eff)!=self.a/2) | (abs(y_eff)!=self.a/2) | (abs(z_eff)!=self.a/2))
-            return idx
-        
-        else: 
-            idx = np.where((abs(x_eff) >= self.a/2) | (abs(y_eff) >= self.a/2) | 
-            (abs(z_eff) >= self.a/2) | ((abs(x_eff) <= self.b/2) 
-            & (abs(y_eff) <= self.b/2) & (abs(z_eff) <= self.b/2)))
-
-        return idx
-
-
-class Cuboid:
-    def __init__(self, dimensions: List[float]):
-        self.a = dimensions[0]
-        self.b = dimensions[1]
-        self.c = dimensions[2]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a cuboid"""
-        return self.a * self.b * self.c
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a cuboid"""
-        x_add = np.random.uniform(-self.a, self.a, Npoints)
-        y_add = np.random.uniform(-self.b, self.b, Npoints)
-        z_add = np.random.uniform(-self.c, self.c, Npoints)
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a Cuboid"""
-        idx = np.where((abs(x_eff) >= self.a/2) 
-        | (abs(y_eff) >= self.b/2) | (abs(z_eff) >= self.c/2))
-        return idx
-
-
-class CylinderRing:
-    def __init__(self, dimensions: List[float]):
-        self.R = dimensions[0]
-        self.r = dimensions[1]
-        self.l = dimensions[2]
-
-    def getVolume(self) -> float:
-        """Returns the volume of a cylinder ring"""
-
-        if self.r > self.R:
-            self.R, self.r = self.r, self.R
-
-        if self.r == self.R:
-            return 2 * np.pi * self.R * self.l #surface area of a cylinder
-
-        else: 
-            return np.pi * (self.R**2 - self.r**2) * self.l
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a cylinder ring"""
-        Volume = self.getVolume()
-
-        if self.r == self.R:
-            #The cylinder ring is a shell
-            phi = np.random.uniform(0, 2 * np.pi, Npoints)
-            x_add = self.R * np.cos(phi)
-            y_add = self.R * np.sin(phi)
-            z_add = np.random.uniform(-self.l / 2, self.l / 2, Npoints)
-            return x_add, y_add, z_add
-
-        Volume_max = 2 * self.R * 2 * self.R * self.l
-        Vratio = Volume_max / Volume
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.R, self.R, N)
-        y = np.random.uniform(-self.R, self.R, N)
-        z = np.random.uniform(-self.l / 2, self.l / 2, N)
-        d = np.sqrt(x**2 + y**2)
-        idx = np.where((d < self.R) & (d > self.r))
-        x_add, y_add, z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a cylinder ring"""
-        d = np.sqrt(x_eff**2 + y_eff**2)
-        if self.r > self.R:
-            self.R, self.r = self.r, self.R
-
-        if self.r == self.R:
-            idx = np.where((d != self.R) | (abs(z_eff) > self.l / 2))
-            return idx
-        else: 
-            idx = np.where((d > self.R) | (d < self.r) | (abs(z_eff) > self.l / 2))
-            return idx
-
-
-class DiscRing(CylinderRing):
-    pass
-
-
-class Superellipsoid:
-    def __init__(self, dimensions: List[float]):
-        self.R = dimensions[0]
-        self.eps = dimensions[1]
-        self.t = dimensions[2]
-        self.s = dimensions[3]
-
-    @staticmethod
-    def beta(a, b) -> float:
-        """beta function"""
-
-        return gamma(a) * gamma(b) / gamma(a + b)
-
-    def getVolume(self) -> float:
-        """Returns the volume of a superellipsoid"""
-
-        return (8 / (3 * self.t * self.s) * self.R**3 * self.eps * 
-                self.beta(1 / self.s, 1 / self.s) * self.beta(2 / self.t, 1 / self.t))
-
-    def getPointDistribution(self, Npoints: int) -> Vector3D:
-        """Returns the point distribution of a superellipsoid"""
-        Volume = self.getVolume()
-        Volume_max = 2 * self.R * self.eps * 2 * self.R * 2 * self.R
-        Vratio = Volume_max / Volume
-
-        N = int(Vratio * Npoints)
-        x = np.random.uniform(-self.R, self.R, N)
-        y = np.random.uniform(-self.R, self.R, N)
-        z = np.random.uniform(-self.R * self.eps, self.R * self.eps, N)
-
-        d = ((np.abs(x)**self.s + np.abs(y)**self.s)**(self.t/ self.s) 
-            + np.abs(z / self.eps)**self.t)
-        idx = np.where(d < np.abs(self.R)**self.t)
-        x_add, y_add, z_add = x[idx], y[idx], z[idx]
-
-        return x_add, y_add, z_add
-
-    def checkOverlap(self, x_eff: np.ndarray, 
-                           y_eff: np.ndarray, 
-                           z_eff: np.ndarray) -> np.ndarray:
-        """Check for points within a superellipsoid"""
-        d = ((np.abs(x_eff)**self.s + np.abs(y_eff)**self.s)**(self.t / self.s) 
-        + np.abs(z_eff / self.eps)**self.t)
-        idx = np.where(d >= np.abs(self.R)**self.t)
-
-        return idx
-
 
 class Qsampling:
     def onQsampling(qmin: float, qmax: float, Nq: int) -> np.ndarray:
@@ -687,7 +192,7 @@ class GenerateAllPoints:
                 "disc_ring": DiscRing,
                 "Disc ring": DiscRing,
                 
-                "superellipsoid": Superellipsoid}
+                "superellipsoid": SuperEllipsoid}
 
     def getSubunitClass(self, key: str):
         if key in self.subunitClasses:
