+++ src/sas/sascalc/shape2sas/models/Disc.py
@@ -1,0 +1,4 @@
+from sas.sascalc.shape2sas.models.EllipticalCylinder import EllipticalCylinder
+
+class Disc(EllipticalCylinder):
+    pass
