+++ src/sas/sascalc/shape2sas/models/DiscRing.py
@@ -1,0 +1,4 @@
+from sas.sascalc.shape2sas.models.CylinderRing import CylinderRing
+
+class DiscRing(CylinderRing):
+    pass
