+++ src/sas/sascalc/shape2sas/models/__init__.py
@@ -1,0 +1,18 @@
+from .Cube import Cube
+from .Cuboid import Cuboid
+from .Cylinder import Cylinder
+from .CylinderRing import CylinderRing
+from .Disc import Disc
+from .DiscRing import DiscRing
+from .Ellipsoid import Ellipsoid
+from .EllipticalCylinder import EllipticalCylinder
+from .HollowCube import HollowCube
+from .HollowSphere import HollowSphere
+from .Sphere import Sphere
+from .SuperEllipsoid import SuperEllipsoid
+
+__all__ = [
+    'Cube', 'Cuboid', 'Cylinder', 'CylinderRing',
+    'Disc', 'DiscRing', 'Ellipsoid', 'EllipticalCylinder',
+    'HollowCube', 'HollowSphere', 'Sphere', 'SuperEllipsoid'
+]
